python/requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.31.0
tqdm>=4.66.0
